requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-debug-toolbar==3.2.2
django-dotenv==1.4.2
django-ratelimit==3.0.1
django-shortuuidfield==0.1.3
django-tinymce4-lite==1.8.0
et-xmlfile==1.1.0
idna==3.2
jsmin==2.2.2
numpy==1.21.2
oauthlib==3.1.1
openpyxl==3.0.7
pandas==1.3.2
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
xlrd==2.0.1

PyMySQL~=1.0.2
lxml~=4.6.3